Create Rectangular_Waveguide_Simulation.m
A script to simulate the behaviour of electric and magnetic fields when interacting with a rectangular waveguide in the dominant mode (TE10). Their behaviour is simulated through plots of the field lines for each respective field propagating relative to cross-sections of the waveguide (front, top, and side views). Based on code from: https://www.mathworks.com/matlabcentral/fileexchange/73101-rectangular-waveguide-e-h-field-plot-for-all-modes.